[step1] 스토리북 마무리
