+++ src/common/components/card/BigCard.tsx
@@ -1,4 +1,4 @@
-import { CardInfo } from "common/types/card.type";
+import { CardInfo } from "features/card/types/card.type";
 
 interface BigCardProps
   extends Pick<CardInfo, "cardNumber" | "cardOwner" | "expireDate"> {}
