+++ src/features/card/utils/inputFormat.ts
