+++ src/pages/SuccessAddCard.tsx
@@ -2,7 +2,7 @@ import Button from "common/components/button/Button";
 import Card from "common/components/card";
 import Title from "common/components/title/Title";
 import Layout from "common/components/layout";
-import { CardInfo } from "common/types/card.type";
+import { CardInfo } from "features/card/types/card.type";
 import { CurrentPage } from "common/types/page.type";
 
 interface SuccessAddCardProps extends CardInfo {
